LPS-28999 Reset organizations LayoutSet when Site delete
Staging turnoff didn't remove staging group.
Deletion of sites belonging to an organization left layoutsets in an inconsistent state.

+++ portal-impl/src/com/liferay/portal/service/impl/GroupLocalServiceImpl.java
@@ -592,18 +592,26 @@ public class GroupLocalServiceImpl extends GroupLocalServiceBaseImpl {
 
 		ServiceContext serviceContext = new ServiceContext();
 
-		try {
-			layoutSetLocalService.deleteLayoutSet(
-				group.getGroupId(), true, serviceContext);
-		}
-		catch (NoSuchLayoutSetException nslse) {
-		}
+		if (!group.isStagingGroup() &&
+			group.isOrganization() && group.isSite()) {
 
-		try {
-			layoutSetLocalService.deleteLayoutSet(
-				group.getGroupId(), false, serviceContext);
+			reInitializeLayout(group.getGroupId(), true, serviceContext);
+			reInitializeLayout(group.getGroupId(), false, serviceContext);
 		}
-		catch (NoSuchLayoutSetException nslse) {
+		else {
+			try {
+				layoutSetLocalService.deleteLayoutSet(
+					group.getGroupId(), true, serviceContext);
+			}
+			catch (NoSuchLayoutSetException nslse) {
+			}
+
+			try {
+				layoutSetLocalService.deleteLayoutSet(
+					group.getGroupId(), false, serviceContext);
+			}
+			catch (NoSuchLayoutSetException nslse) {
+			}
 		}
 
 		// Group roles
@@ -635,11 +643,21 @@ public class GroupLocalServiceImpl extends GroupLocalServiceBaseImpl {
 		teamLocalService.deleteTeams(group.getGroupId());
 
 		// Staging
-
 		unscheduleStaging(group);
 
 		if (group.hasStagingGroup()) {
-			deleteGroup(group.getStagingGroup().getGroupId());
+
+
+			try {
+				StagingUtil.disableStaging(group, group, serviceContext);
+
+				//deleteGroup(group.getStagingGroup().getGroupId());
+			}
+			catch (Exception e) {
+				if (_log.isErrorEnabled()) {
+					_log.error("Unable to disable staging for group: " + group);
+				}
+			}
 		}
 
 		// Themes
@@ -743,7 +761,9 @@ public class GroupLocalServiceImpl extends GroupLocalServiceBaseImpl {
 
 		// Group
 
-		if (group.isOrganization() && group.isSite()) {
+		if (!group.isStagingGroup() &&
+			group.isOrganization() && group.isSite()) {
+
 			group.setSite(false);
 
 			groupPersistence.update(group, false);
@@ -2676,6 +2696,48 @@ public class GroupLocalServiceImpl extends GroupLocalServiceBaseImpl {
 			if (name.equals(company.getName())) {
 				throw new DuplicateGroupException();
 			}
+		}
+	}
+
+	/**
+	 * Deletes a layoutset and creates a new one, and keeps it's logo if it had
+	 * one.
+	 * @param groupId - Id of the group the layoutset belongs to
+	 * @param privateLayoutSet - Determines if it's a private layoutset
+	 * @param serviceContext - The service context
+	 * @throws PortalException
+	 * @throws SystemException
+	 */
+	public void reInitializeLayout(long groupId, boolean privateLayoutSet,
+		ServiceContext serviceContext)
+		throws PortalException, SystemException
+	{
+		try {
+			LayoutSet layoutSet = layoutSetLocalService.getLayoutSet(groupId,
+				privateLayoutSet);
+			boolean logo = layoutSet.getLogo();
+			long logoId = layoutSet.getLogoId();
+
+			if (logo) {
+				layoutSet.setLogo(false);
+				layoutSet.setLogoId(0);
+				layoutSetPersistence.update(layoutSet, false);
+			}
+
+			layoutSetLocalService.deleteLayoutSet(groupId, privateLayoutSet,
+				serviceContext);
+
+			layoutSet = layoutSetLocalService.addLayoutSet(groupId,
+				privateLayoutSet);
+
+			if (logo) {
+				layoutSet.setLogo(logo);
+				layoutSet.setLogoId(logoId);
+				layoutSetPersistence.update(layoutSet, false);
+			}
+		}
+
+		catch(NoSuchLayoutSetException nsle) {
 		}
 	}
